requirements.txt
numpy==1.23
pandas==2.2.2
python-dateutil==2.8.2
pytz==2021.1
scipy==1.10.1
folium==0.13.0
streamlit-folium==0.20.0
python-dotenv==1.0.1
streamlit== 1.34.0

